widen ParseInt calls to accept too-big codepoints
